fixing the workflow so psydac runs
